pyproject.toml
[project]
name = "dedup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest>=8.4.2",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "trafilatura>=2.0.0",
    "xxhash>=3.6.0",
]
